utils: extracted ModuleCache from resolve cache behavior
