+++ utils/package.json
@@ -1,6 +1,6 @@
 {
   "name": "eslint-module-utils",
-  "version": "0.2.0",
+  "version": "0.2.1",
   "description": "Core utilities to support eslint-plugin-import and other module-related plugins.",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
